static sp
